refactor recipe pack tests into seperate file
